Add an initial implementation of UTF-8 and bytestring interners
SymbolTable wraps a Vec and HashMap and wraps many of their APIs for managing
the data structure.

SymbolTable has several APIs that yield iterators for inspecting the contents
of the table.

The primary interface is SymbolTable::intern, which takes owned or borrowed strings
and returns u32 tokens, and SymbolTable::get, which takes u32 tokens and looks up
the underlying string in the table.

SymbolTable is implemented twice, once for byteslices, &[u8], and once for UTF-8
string slices, &str. The two structs have nearly identical interfaces.