Reorganize public API to make it easier to add cargo features
Intaglio's only external dependencies are used by the bytestring
interner, so invert the crate organization to make the default exported
interner the UTF-8 string interner. This makes it easier to hide `bytes`
in a module that can be disabled.

This commit includes many other changes to prep for release:

- Fixup Cargo.toml description metadata to mention both UTF-8 and
  bytestring support.
- Fixup Cargo.toml repository metadata to point at the correct
  repository since this was broken out of artichoke/artichoke.
- Add `utf8` Cargo.toml keyword.

Add an optional `bytes` feature that is enabled by default. Enabling
this feature includes a `SymbolTable` that interns binary data.
Disabling this feature drops the `bstr` dependency.

Add documentation in `Cargo.toml` on the crate features with inline
comments.

Rework README to shift emphasis to UTF-8 string interner. Document
`bytes` crate feature in README.

Fixup benches to run when passed `--no-default-features`.

Fixup doctests to run when passed `--no-default-features`.

Rename `SymbolId` to `Symbol`. This undoes a hack to avoid name
collisions in Artichoke where `Symbol` refers to the core type and mruby
uses `sym_id` as the typedef for the identifiers boxed into the
`mrb_value`. This also caused a rename of `SymbolIdOverflowError` to
`SymbolOverflowError`.

Rework crate-level and module-level documentation to account for the
`bytes` module being optional. I used `regexp` and `regexp::bytes` as
inspiration here. This massively improves differentition between the
`str` and `bytes` interners.

Re-export crate-level types and constants, e.g. `Symbol`,
`SymbolOverflowError`, and `DEFAULT_SYMBOL_TABLE_CAPACITY` from `bytes`
module.

Restructure the crate-level docs to have this structure:

> this crate is a string interner.
> this api is like a bimap.
> refer to `SymbolTable` for more specific API documentation.
> example
> what is a string interner
> describe allocation behavior and knobs
> enumerate and describe cargo featutres