Prevent mutable access to the owned contents of the Slice enum
See GH-29. Although the Box in Slice::Owned is the unique owner of the slice
contents, we are handling out static references to the interior, which requires
its contents not be moved out.

Enforce this invariant by wrapping the Slice enum in an opaque struct that disallows
mutable access to the interior. Module visibility enforces this as well.

This commit also removes BStr from the bytes interner, replacing its debug impl with
String::from_utf8_lossy. The bstr crate is removed. Intaglio now has zero runtime
dependencies, other than std.