Benchmark `bytes::SymbolTable`
Prior to #23, the criterion benchmarks were defined in the `intaglio`
crate which means the cfgs for `feature = "bytes"` were necessary to
bench with no default features.

In #23, the benchmarks suite was copied over without change. This
oversight means that the `bytes` benchmarks have never run since there
is no `bytes` feature in the `benchmarks` crate.

Remove the `cfg` pragmas and fixup the benchmark groups for the bytes
tables so they are distinguishable from the `str::SymbolTable` tests in
the criterion output.