Simplify and shrink `AllSymbols` iterators
Remove the "either" branching of the inner range in the `AllSymbols`
iterators.

Use a `Range<usize>` rather than converting to either a `Range<u32>` or
`RangeInclusive<u32>`. Do the case to u32 -> Symbol when advancing the
iterator.

This removes a bunch of branching and reduces the size of the struct.