Srink `AllSymbols` iterators, fix max `SymbolTable` size
The changes in #139 to shrink `AllSymbols` actually fixed a bug where
the iterator included `u32::MAX` as a symbol. The intent of the
`SymbolTable`s is to hold _at most_ `u32::MAX` symbols. Because sym ids
start at zero, that means the maximum sym id is `u32::MAX - 1`.

This invariant allows the `Range` in `AllSymbols` to be a `Range<u32>`
because `(0..u32::MAX).count() == u32::MAX`.

The possible range of `(0..=u32::MAX).count() == u32::MAX + 1`, which
means the symbol table would be storing more symbols than intended.