Address stacked borrows violation in `SymbolTable::intern`
This stacked borrows violation impacts all `SymbolTable` variants in
this crate and has been present in all published versions.

The `SymbolTable::intern` function previously worked like this:

1. Convert the given contents to be interned into a `Cow<'static, T>`.
2. Check to see if this content is already interned by checking the
  internal bimap. If found, return the matching symbol.
3. Convert the `Cow` into an `Interned` which is an enum of either a
  &'static T` or a `Box<T>`.
4. Borrow a `&'static T` from the `Interned`.
5. Push the `Interned` into the vec.
6. Add the `&'static T` to the map.
7. Check that the insertion happened correctly with some debug
  assertions.
8. Return the computed symbol.

As of at least rustc 1.73.0-nightly (31395ec38 2023-07-24), Miri pops
all borrows from the `Interned` at step 5 when the `Interned` is moved
into the vec, which means the borrow made at 4 is untracked when Miri
executes the debug assertions.

The fix here is to make the assignment/move of the `Interned` happen
first so Miri doesn't retag the inner `Box<T>`. This is accomplished by
reordering the `self.vec.push` operation to immediately follow step 3.
An additional unchecked index into the vec is used to get a shared
access to that `Interned` object to later derive the `&'static T`.

This commit also adds an additional test to resolve a slice from the
symbol table to ensure Miri does not disagree as the `Interned` is moved
during growth of the `SymbolTable`.

Fixes #235.