Remove `version-sync` and version wholesome tests
I use a tool called `version-sync` to run wholesome tests on
`Cargo.toml`, `README.md`, and `src/lib.rs` to make sure the
version of the crate in `Cargo.toml` matches in all of these
places.

`version-sync` depends on the `toml` crate to do TOML parsing
of "add to cargo.toml" dependency snippets.

The toml crate has undergone a lot of churn under its new maintainers.

The repository for the `toml` crate has rennovate bot set up to update
the MSRV of the `toml` crate as declared in Cargo.toml automatically.
These MSRV metadata updates are published in patch releases.

Cargo will download dependencies and do MSRV resolution even if the deps
are cfg'd out.

This means either the `toml` crate is deciding my MSRV for me or I cannot
use `version-sync` crate for wholesome tests.

It is not possible to move to `basic-toml`, which is a dtolnay fork of the
`toml` v0.5 branch. `basic-toml` doesn't support parsing directly into a
`Value`, only going through `serde`.